Delete function Update

+++ src/static/script.js
@@ -3,19 +3,32 @@ document.addEventListener("DOMContentLoaded", function() {
     var deleteTaskButton = document.getElementById('deleteButton');
     var taskInput = document.getElementById('taskInput');
     var taskList = document.getElementById('taskList');
-
+	let item = null;
     async function addTask() {
         var taskText = taskInput.value.trim();
         if (taskText === '') return;
 
         var listItem = document.createElement('li');
         listItem.textContent = taskText;
+		listItem.dataset.id = taskList.children.length - 1;
+		listItem.dataset.title = taskText;
 
         // Add event listeners for both click and touch events
-        listItem.addEventListener('click', toggleTaskCompletion);
+        listItem.addEventListener('click', function(event) {
+			toggleTaskCompletion.call(this);
+			event.preventDefault();
+			if(null == item) {
+				item = event.target;
+				console.log(item.dataset.id);
+			} else{
+				item = null;
+			}
+		});
         listItem.addEventListener('touchend', function(event) {
             toggleTaskCompletion.call(this);
             event.preventDefault(); // Prevents click event from being triggered after touchend
+			item = event.target;
+			alert(item);
         });
 		// async + await = 비동기 코드를 동기 코드처럼 작성할 수 있어, 가독성 좋아지고 에러 처리가 간단
 		// Promise 는 비동기 함수의 결과를 담고 있는 객체
@@ -24,7 +37,7 @@ document.addEventListener("DOMContentLoaded", function() {
 		async : 해당 함수는 항상 프로미스를 반환하도록 지정 (이 함수는 비동기적인 함수이고 Promise 를 반환)
 		await : 프로미스가 처리되거나 거불될 때까지 실행을 일시 중지
 		*/
-		response = await fetch('http://127.0.0.1:8000/guestbook', {
+		response = await fetch('http://172.30.1.53:8000/guestbook', {
 		method: 'POST',
 		headers: {
 			'accept': 'application/json',
@@ -47,15 +60,18 @@ document.addEventListener("DOMContentLoaded", function() {
 
 
     async function deleteSelectedItem() {
-        id = 0;
-        const response = await fetch('http://127.0.0.1:8000/guestbook', {
+		if(item == null) {
+			alert("Select Item!!!");
+			return;
+		}
+        const response = await fetch('http://172.30.1.53:8000/guestbook', {
             method: 'DELETE',
             headers: {
                 'Content-Type': 'application/json'
             },
             body: JSON.stringify({
-                id:-1,
-                title:"aaa"
+                id: item.dataset.id,
+                title: item.dataset.title
             })
         });
         if(response.ok) {
@@ -63,6 +79,7 @@ document.addEventListener("DOMContentLoaded", function() {
         } else {
             alert('Failed to delete item');
         }
+		item = null;
         showTodos();
     }
 
@@ -79,17 +96,31 @@ document.addEventListener("DOMContentLoaded", function() {
     deleteTaskButton.addEventListener('click', deleteSelectedItem);
 
 	async function showTodos() {
-		const response = await fetch('http://127.0.0.1:8000/guestbook/all');
+		const response = await fetch('http://172.30.1.53:8000/guestbook/all');
 		const todos = await response.json();
 		const todoList = document.getElementById("taskList");
 		todoList.classList.add("custom-list");
 		todos.forEach(todo => {
 			const listTodo = document.createElement("li");
-			listTodo.addEventListener("click", toggleTaskCompletion)
-			listTodo.addEventListener('touchend', function(event) {
-        	    toggleTaskCompletion.call(this);
-            	event.preventDefault(); // Prevents click event from being triggered after touchend
-	        });
+			listItem.dataset.id = taskList.children.length - 1;
+			listItem.dataset.title = taskText;
+			// Add event listeners for both click and touch events
+        listTodo.addEventListener('click', function(event) {
+			toggleTaskCompletion.call(this);
+			event.preventDefault();
+			if(null == item) {
+				item = event.target;
+				console.log(item.dataset.id);
+			} else{
+				item = null;
+			}
+		});
+        listTodo.addEventListener('touchend', function(event) {
+            toggleTaskCompletion.call(this);
+            event.preventDefault(); // Prevents click event from being triggered after touchend
+			item = event.target;
+			alert(item);
+        });
 			listTodo.textContent = todo.title;
 			todoList.appendChild(listTodo);
 		});
